refactor: rename WalletButton to LoginButton
