+++ web-marketplace/src/components/organisms/LoginButton/LoginButton.types.ts
@@ -1,0 +1,1 @@
+export type ButtonSize = 'small' | 'medium' | 'large';
