--- web-marketplace/src/components/organisms/WalletButton/WalletButton.types.ts
@@ -1,7 +1,0 @@
-export type WalletUiItem = {
-  name: string;
-  description: string;
-  imageUrl: string;
-};
-
-export type ButtonSize = 'small' | 'medium' | 'large';
